added a bunch of stuff
nobody is ever going to read this so i may as well write whatever the hell i feel like in here. one of my biggest desires in this life is to be duct taped to a ceiling. i'm serious, i mean it. i will literally pay you to do that. only condition, of course, is that you have to un-duct-tape me soon enough. oh and another one of my burning desires in life is that i really want to get blackout drunk and wake up, alone, in the middle of one of those escape puzzle room thingies, with no possibility of a return to civilisation before i solve the puzzle, and (this is important) no idea who it was who put me in the room. just make sure you give me a toilet as well